refactor(mux/provider): moved queue logic into new queue.ts file
